+++ app/controllers/images_controller.rb
@@ -186,12 +186,12 @@ class ImagesController < ApplicationController
 
   # GET 'images/:id/scale_to_box/:x/:y/:width/:height/:box_width/:box_height'
   def scale_to_box
-    send_data Image.scaled_to_box_blob(params), type: 'image/png', disposition: 'inline'
+    send_data Image.scaled_to_box_blob(params), type: 'image/jpg', disposition: 'inline'
   end
 
   # GET 'api/v1/images/:id/scale_to_box/:x/:y/:width/:height/:box_width/:box_height'
   def api_scale_to_box
-    send_data Image.scaled_to_box_blob(params), type: 'image/png', disposition: 'inline'
+    send_data Image.scaled_to_box_blob(params), type: 'image/jpg', disposition: 'inline'
   end
 
   # GET 'images/:id/as_png'