+++ app/javascript/vue/components/radials/co/components/SliceAccessions.vue
@@ -60,6 +60,7 @@ import { computed, ref } from 'vue'
 import { CollectionObject } from '@/routes/endpoints'
 import PreviewBatch from '@/components/radials/shared/PreviewBatch.vue'
 import UpdateBatch from '@/components/radials/shared/UpdateBatch.vue'
+import updateMessage from '../utils/updateMessage.js'
 
 const MAX_LIMIT = 250
 
@@ -94,12 +95,4 @@ const payload = computed(() => ({
     ...data.value
   }
 }))
-
-function updateMessage(data) {
-  const message = data.sync
-    ? `${data.updated.length} collection objects queued for updating.`
-    : `${data.updated.length} collection objects were successfully updated.`
-
-  TW.workbench.alert.create(message, 'notice')
-}
 </script>