+++ spec/models/source/bibtex_spec.rb
@@ -21,7 +21,7 @@ describe Source::Bibtex, type: :model, group: :sources do
     BibTeX.open(Rails.root + 'spec/files/bibtex/Taenionema.bib')
   }
 
-  specify '.batch_update' do
+  specify '.batch_update (sync)' do
     s1 = FactoryBot.create(:valid_serial)
     s2 = FactoryBot.create(:valid_serial)
 
@@ -37,6 +37,31 @@ describe Source::Bibtex, type: :model, group: :sources do
     expect(response[:updated]).to include(sb.id)
     expect(response[:not_updated]).to eq([])
     expect(sb.reload.serial).to eq s2
+  end
+
+  specify '.batch_update (async)' do
+    s1 = FactoryBot.create(:valid_serial)
+    s2 = FactoryBot.create(:valid_serial)
+
+    sb1 = FactoryBot.create(:valid_source_bibtex, serial: s1)
+    sb2 = FactoryBot.create(:valid_source_bibtex, serial: s1)
+
+    params = {
+      async_cutoff: 1,
+      source: { serial_id: s2.id },
+      user_id: Current.user_id,
+      project_id: Current.project_id
+    }.merge(source_query: {source_id: [sb1.id, sb2.id]})
+
+    response = Source::Bibtex.batch_update(params).to_json
+
+    sleep(2) # jobs trigger in 1 second
+    Delayed::Worker.new.work_off
+
+    expect(response[:total_attempted]).to eq(2)
+    expect(response[:async]).to eq(true)
+    expect(sb1.reload.serial).to eq s2
+    expect(sb2.reload.serial).to eq s2
   end
 
   specify '#project_sources_attributes 3' do
@@ -153,6 +178,50 @@ describe Source::Bibtex, type: :model, group: :sources do
     src = Source::Bibtex.new_from_bibtex_text(citation_string)
     src.save!
     expect(src.serial_id).to eq(s.id)
+  end
+
+  specify '.new_from_bibtex without create project source' do
+    citation_string =  %q{@Article{Park2021a,
+        author = {Kyu-Tek Park AND J. B. Heppner},
+        title = {Notes on Vietnam moths, 21. Athymoris in Vietnam, with two new species (Lepidoptera: Lecithoceridae: Torodorinae)},
+        journal = {Lepidoptera Novae},
+        year = {2021},
+        volume = {13},
+        number = {1-2},
+        pages = {23--26},
+        issn = {1941-1014},
+        abstract = {– Two new species of the genus Athymoris Meyrick, 1935, are described from Vietnam: Athymoris gilvimaculata Park & Heppner, sp. nov.,
+        and A. clinozonalis Park & Heppner, sp. nov.},
+        file = {:VN-Athymoris.pdf:PDF},
+        }}
+
+    src = Source::Bibtex.new_from_bibtex_text(citation_string)
+    src.save!
+
+    project_source = ProjectSource.find_by(source_id: src.id, project_id: Current.project_id)
+    expect(project_source).to be_nil
+  end
+
+  specify '.new_from_bibtex create project source' do
+    citation_string =  %q{@Article{Park2021a,
+        author = {Kyu-Tek Park AND J. B. Heppner},
+        title = {Notes on Vietnam moths, 21. Athymoris in Vietnam, with two new species (Lepidoptera: Lecithoceridae: Torodorinae)},
+        journal = {Lepidoptera Novae},
+        year = {2021},
+        volume = {13},
+        number = {1-2},
+        pages = {23--26},
+        issn = {1941-1014},
+        abstract = {– Two new species of the genus Athymoris Meyrick, 1935, are described from Vietnam: Athymoris gilvimaculata Park & Heppner, sp. nov.,
+        and A. clinozonalis Park & Heppner, sp. nov.},
+        file = {:VN-Athymoris.pdf:PDF},
+        }}
+
+    src = Source::Bibtex.new_from_bibtex_text(citation_string, Current.project_id)
+    src.save!
+
+    project_source = ProjectSource.find_by(source_id: src.id, project_id: Current.project_id)
+    expect(project_source).to be_a(ProjectSource)
   end
 
   specify '#year_with_suffix' do